+++ database/migrations/20240929005504_budget.sql
@@ -6,7 +6,7 @@ CREATE TABLE
     updated_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
     name VARCHAR(255) NOT NULL,
     income_or_expense TEXT CHECK (income_or_expense IN ('income', 'expense')) NOT NULL,
-    amount DECIMAL(10, 2) NOT NULL
+    amount INTEGER NOT NULL
   );
 
 CREATE TABLE
@@ -17,7 +17,7 @@ CREATE TABLE
     transaction_date DATE NOT NULL,
     category_id UUID NOT NULL,
     description TEXT NOT NULL,
-    amount DECIMAL(10, 2) NOT NULL,
+    amount INTEGER NOT NULL,
     FOREIGN KEY (category_id) REFERENCES categories (id)
   );
